Add userid-header parameter to profiles (#1700)
Without this commit, the profiles controller will have a "userid-header"
parameter of "" set, which cause the Profile CRD to create an Istio
ServiceRoleBinding with an incorrect subject "request.headers[]:
<profile username>" which will never match anything and cause the user
to not be able to connect to their Jupyter notebooks. This commit
ensures the profiles controller is configured correctly to fix this
problem.